requirements_dev.txt
-e .
check-manifest
pytest-flake8
readme_renderer[md]
setuptools>=38.6.0
twine>=1.13.0
viscm>=0.9
wheel>=0.31.0
